requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
Django==4.0
django-phonenumber-field==6.0.0
pycparser==2.21
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
